[obexd] Prevent double disconnect and null reference. Contributes to JB#9622.
Previous patch did not consider that obc_session_shutdown() can be
called for more than one reason: it is called when the D-Bus client
that requested the OBEX session unregisters, but also when it leaves
D-Bus. This was not a problem before, but now that obex-client waits
for disconnection reply it may see both events.

Keeping track when disconnection is being executed with a flag
prevents trying to do it twice.

Also, don't send a disconnect request if we don't have
a valid OBEX handle.